requirements.txt
Django==1.6.5
#PIL==1.1.7
Pillow
South==1.0
django-crispy-forms==1.4.0
django-ckeditor==4.0.2
Markdown==2.4
django-contrib-comments
